pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-bridge"
version = "0.1.0"
description = "Model Context Protocol bridge server for ospsuite"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "PBPK MCP Team" }
]
license = { text = "Proprietary" }
keywords = ["mcp", "ospsuite", "fastapi"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Framework :: FastAPI",
  "License :: Other/Proprietary License"
]
dependencies = [
  "fastapi>=0.110,<1.0",
  "uvicorn[standard]>=0.29,<1.0",
  "pydantic>=2.6,<3.0",
  "python-dotenv>=1.0,<2.0",
  "structlog>=24.1,<25.0",
  "langchain-core>=0.2.10,<0.3",
  "langchain-community>=0.2.6,<0.3",
  "langgraph>=0.0.50,<0.1",
  "python-jose[cryptography]>=3.3,<4.0",
  "httpx>=0.27,<0.29"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2,<9.0",
  "pytest-asyncio>=0.23,<1.0",
  "httpx>=0.27,<0.28",
  "ruff>=0.6.4,<0.7.0",
  "black>=24.4,<25.0",
  "mypy>=1.10,<2.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008", "UP006", "UP007", "UP035"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
  "ignore::DeprecationWarning"
]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.9"
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
strict_optional = true
packages = ["mcp_bridge"]
exclude = ["tests"]
[[tool.mypy.overrides]]
module = [
  "langchain_core.*",
  "langchain_community.*",
  "langgraph.*"
]
ignore_missing_imports = true
